requirements.txt
selenium==4.15.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.0
fake-useragent==1.4.0
webdriver-manager==4.0.1